fix(trusty-mpm): self-guard graceful_terminate_runtime when pane already gone (review)
Addresses the code-review HIGH finding on #1975: stop() called
graceful_terminate_runtime unconditionally, so stopping a session whose tmux
pane was already gone paid the full SIGTERM_GRACE_SECS sleep before kill_session
failed — an N×grace slowdown for reconciler/prune-idle loops over dead sessions.

Push the session_exists fast-path INTO graceful_terminate_runtime (early return,
no signal/sleep/kill) so both stop() and decommission benefit; drop the now
redundant session_exists guard in decommission_with_root. Add
graceful_terminate_runtime_noop_when_session_gone; relocate both
graceful_terminate_runtime tests from the at-cap tests.rs to restart_tests.rs.
